renames use get history hook file to match repo conventions
